move observer to trainer module
