+++ trainer/__init__.py
@@ -1,2 +1,3 @@
 from trainer.config import *
 from trainer.train import *
+import trainer.observers as observers
